fix: export only active consensus validators to genesis
When exporting a chain to a genesis file, there's a field in the genesis
called "validators" that contains the validators that'll be used
to bootstrap consensus. This is separate to the list of all validators,
bonded and otherwise, kept in the staking module's app_state. If this field
contains anything other than those validators that're actively
participating in consensus, the chain import will fail. Thus,
we shouldn't add all the validators from the staking module to this
field, but only those we want used for consensus.